Refactor various controller into one decisions controller
